New Attack system deals damage
