+++ Assets/Scripts/AttackAreaOfEffect.cs
@@ -1,8 +1,9 @@
 ﻿using System.Collections;
+using AttackList;
 using UnityEngine;
 
 
 public abstract class AttackAreaOfEffect : MonoBehaviour
 {
-    public abstract IEnumerator Attack(Vector2 direction, Transform[] exitPoints, GameObject spriteObject);
+    public abstract IEnumerator Attack(Vector2 direction, Transform[] exitPoints,AttackBase attack);
 }
